+++ src/tdb_decode.c
@@ -64,7 +64,7 @@ int tdb_get_trail_filtered(const tdb *db,
                            const uint32_t *filter,
                            uint32_t filter_len)
 {
-    static const int INITIAL_ITEMS_BUF_LEN = 1 << 16;
+    static const uint32_t INITIAL_ITEMS_BUF_LEN = 1U << 16;
 
     if (!*items_buf_len){
         if (!(*items = malloc(INITIAL_ITEMS_BUF_LEN * 4)))
@@ -148,7 +148,7 @@ uint32_t tdb_decode_trail_filtered(const tdb *db,
         if (delta == TDB_FAR_TIMEDELTA){
             dst[i++] = TDB_FAR_TIMESTAMP;
         }else{
-            tstamp += delta;
+            tstamp += (uint32_t)delta;
             dst[i++] = tstamp;
         }
         item >>= 32;
@@ -156,9 +156,9 @@ uint32_t tdb_decode_trail_filtered(const tdb *db,
         /* handle a possible latter part of the first bigram */
         if (item){
             field = tdb_item_field(item);
-            db->previous_items[field] = item;
+            db->previous_items[field] = (uint32_t)item;
             if (edge_encoded && i < dst_size)
-                dst[i++] = item;
+                dst[i++] = (uint32_t)item;
         }
 
         /* decode one event: timestamp is followed by at most num_ofields
